added workaround for redirection problems
added copy link for git pages because links don't work in the embedded page we have on the Ocean Navigator. 
If you have a hard time opening links on this page please try pasting https://dfo-ocean-navigator.github.io/Ocean-Navigator-Manual/ into your browser or right click on the link and select open in a new tab.